Manage Listing Page & User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -15,7 +15,7 @@ class ListingController extends Controller
         ]);
     }
 
-    // Show single listing
+    //Show single listing
     public function show(Listing $listing) {
         return view('listings.show', [
             'listing' => $listing
@@ -31,7 +31,7 @@ class ListingController extends Controller
     public function store(Request $request) {
         $formFields = $request->validate([
             'title' => 'required',
-            'company' => ['required', Rule::unique('listings','company')],
+            'company' => ['required', Rule::unique('listings', 'company')],
             'location' => 'required',
             'website' => 'required',
             'email' => ['required', 'email'],
@@ -57,6 +57,11 @@ class ListingController extends Controller
 
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -78,7 +83,17 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
-        return redirect('/')->with('message', 'Listing deleted successfully!');
+        return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
